Receiving accepts decimal quantities (QA list, item 6)
The Receive Qty / Ordered Qty inputs had no step, so the browser refused
e.g. 12.5; behind that, validation was integer-only, the service
truncated with intval(), and purchase_order_items quantities were
integer columns. Stock tables (product_stocks, product_batches,
stock_transactions) were already numeric.

- Migration: purchase_order_items.requested_quantity / received_quantity
  -> numeric(15,2) (lossless; checked on the live table in a rolled-back
  transaction: 387 rows intact, 2.5 received and stored exactly).
- Model casts both to float so whole numbers still show as "100".
- Request: numeric; service / over-receipt: floats rounded to 2 dp;
  inputs step="0.01"; the landed-cost preview uses parseFloat.
- Tests compare quantities numerically; new decimal receive test.

+++ app/Services/OverReceiptService.php
@@ -74,8 +74,8 @@ class OverReceiptService
                     continue;
                 }
                 $item->requested_quantity = $decision === 'approve'
-                    ? max((int) $item->requested_quantity, (int) $item->received_quantity)
-                    : (int) $line['ordered'];
+                    ? max((float) $item->requested_quantity, (float) $item->received_quantity)
+                    : (float) $line['ordered'];
                 $item->save();
             }
